refactor(tui): extract input handlers into state/keys.rs
Move the user-input half of the reducer into state/keys.rs: the per-mode
key handlers (reduce_key_normal/leader/command/picker/confirm), mouse
reduce_click, command-palette run_command, and the detail-navigation
helpers they call (split_detail, focus_or_solo, close_focused,
cycle_focused_autonomy, close_agent_by_id, request_spawn). The handlers
reduce_inner dispatches to are pub(super); the nav helpers stay private to
keys. mark_shown_seen (shared with the event reducer) stays at the state
root. No behavior change.
